requirements.txt
alabaster==1.0.0
attrs==23.2.0
babel==2.16.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.4.1
colorama==0.4.6
et-xmlfile==1.1.0
greenlet==3.1.1
h11==0.14.0
idna==3.7
imagesize==1.4.1
Jinja2==3.1.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
mysql-connector==2.2.9
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
PyAutoIt==0.6.5
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
pywin32==308
requests==2.32.3
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
stringcase==1.2.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.1
wsproto==1.2.0
xlwings==0.33.5
